requirements.txt
grpcio
protobuf
paho-mqtt
pysyncobj